requirements.txt
annotated-types==0.6.0
colorama==0.4.6
future==0.18.2
guiElements==0.0.6
mando==0.6.4
pygame-ce==2.4.0
PyYAML==6.0.1
radon==5.1.0
six==1.16.0
typing_extensions==4.9.0
